pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "espin"
version = "0.1.0"
description = "Local Streaming ASR for Coding Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx-whisper>=0.4.3",
    "numpy>=1.26.0",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=10.0",
    "pyobjc-framework-Quartz>=12.1",
    "pyobjc-framework-AVFoundation>=10.0",
    "sounddevice>=0.5.5",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["espin*"]
exclude = ["Formula*", "test_*", ".pi*"]

[tool.setuptools]
py-modules = ["espin_gui"]

[project.scripts]
espin = "espin.main:main"
espin-gui = "espin_gui:main"
